refactoring: move OSGi resource provider impl to flow-server
